+++ src/types.ts
@@ -1,0 +1,13 @@
+import { TimeRange } from "fiberplane-charts";
+
+export type AbsoluteTimeRange = {
+  type: "absolute";
+} & TimeRange;
+
+export type RelativeTimeRange = {
+  type: "relative";
+  from: string;
+  to: string;
+};
+
+export type FlexibleTimeRange = AbsoluteTimeRange | RelativeTimeRange;